Moved Json string helper functions to seperate class
